+++ malariagen_data/anoph/karyotype_params.py
@@ -1,10 +1,9 @@
 """Parameter definitions for karyotype analysis functions."""
 
-from typing import Literal
 
 from typing_extensions import Annotated, TypeAlias
 
 inversion_param: TypeAlias = Annotated[
-    Literal["2La", "2Rb", "2Rc_gam", "2Rc_col", "2Rd", "2Rj"],
+    str,
     "Name of inversion to infer karyotype for.",
 ]
