+++ malariagen_data/anopheles.py
@@ -29,7 +29,7 @@ from .anoph import (
     plotly_params,
     xpehh_params,
 )
-from .anoph.karyotype import AnophelesKaryotypeData
+from .anoph.karyotype import AnophelesKaryotypeAnalysis
 from .anoph.aim_data import AnophelesAimData
 from .anoph.base import AnophelesBase
 from .anoph.cnv_data import AnophelesCnvData
@@ -95,7 +95,7 @@ class AnophelesDataResource(
     AnophelesPca,
     PlinkConverter,
     AnophelesIgv,
-    AnophelesKaryotypeData,
+    AnophelesKaryotypeAnalysis,
     AnophelesAimData,
     AnophelesHapData,
     AnophelesSnpData,
