+++ malariagen_data/util.py
@@ -1591,27 +1591,6 @@ def distributed_client():
     return client
 
 
-def _karyotype_tags_n_alt(gt, alts, inversion_alts):
-    # could be Numba'd for speed but was already quick (not many inversion tag snps)
-    n_sites = gt.shape[0]
-    n_samples = gt.shape[1]
-
-    # create empty array
-    inv_n_alt = np.empty((n_sites, n_samples), dtype=np.int8)
-
-    # for every site
-    for i in range(n_sites):
-        # find the index of the correct tag snp allele
-        tagsnp_index = np.where(alts[i] == inversion_alts[i])[0]
-
-        for j in range(n_samples):
-            # count alleles which == tag snp allele and store
-            n_tag_alleles = np.sum(gt[i, j] == tagsnp_index[0])
-            inv_n_alt[i, j] = n_tag_alleles
-
-    return inv_n_alt
-
-
 def prep_samples_for_cohort_grouping(*, df_samples, area_by, period_by):
     # Take a copy, as we will modify the dataframe.
     df_samples = df_samples.copy()
